LayoutTests/platform/win/plugins/get-value-netscape-window-expected.txt

CONSOLE MESSAGE: PLUGIN: NPN_GetValue(NPNVnetscapeWindow) succeeded

This tests that the browser supports NPN_GetValue(NPNVnetscapeWindow).
